Move internal models out of the models folder
