+++ src/zenml/models/v2/misc/build_item.py
@@ -13,12 +13,9 @@
 #  permissions and limitations under the License.
 """Model definition for pipeline build item."""
 
-import hashlib
-from typing import List, Optional, Set
+from typing import Optional
 
 from pydantic import BaseModel, Field
-
-from zenml.config.build_configuration import BuildConfiguration
 
 
 class BuildItem(BaseModel):
@@ -50,60 +47,3 @@ class BuildItem(BaseModel):
     requires_code_download: bool = Field(
         default=False, title="Whether the image needs to download files."
     )
-
-
-class PreparedBuildItem(BaseModel):
-    """A build configuration with its precomputed identifiers.
-
-    Attributes:
-        configuration: The Docker build configuration.
-        key: The key used to identify the image in a pipeline build.
-        settings_checksum: Checksum of the build configuration settings.
-    """
-
-    configuration: BuildConfiguration = Field(
-        title="The Docker build configuration."
-    )
-    key: str = Field(title="The pipeline build image key.")
-    settings_checksum: str = Field(
-        title="The checksum of the build configuration settings."
-    )
-
-
-class PreparedPipelineBuild(BaseModel):
-    """Precomputed data needed to find or create a pipeline build.
-
-    Attributes:
-        items: Build configurations with their precomputed identifiers.
-    """
-
-    items: List[PreparedBuildItem] = Field(
-        title="The prepared Docker build configurations."
-    )
-
-    @property
-    def checksum(self) -> str:
-        """Compute the aggregate checksum of all prepared build items.
-
-        Returns:
-            The aggregate pipeline build checksum.
-        """
-        hash_ = hashlib.md5()  # nosec
-        for item in self.items:
-            hash_.update(item.key.encode())
-            hash_.update(item.settings_checksum.encode())
-
-        return hash_.hexdigest()
-
-    def prune(self, required_keys: Set[str]) -> "PreparedPipelineBuild":
-        """Create a prepared build containing only required build items.
-
-        Args:
-            required_keys: Keys of build items that should be retained.
-
-        Returns:
-            A new prepared pipeline build containing the required items.
-        """
-        return PreparedPipelineBuild(
-            items=[item for item in self.items if item.key in required_keys]
-        )
